fix(net): listen i32

+++ compio-net/src/tcp.rs
@@ -891,12 +891,12 @@ impl TcpSocket {
 
     /// Converts the socket into a `TcpListener`.
     ///
-    /// `backlog` defines the maximum number of pending connections that are queued
-    /// by the operating system at any given time. Connections are removed from
-    /// the queue with [`TcpListener::accept`]. When the queue is full, the
-    /// operating system will start rejecting connections.
-    pub async fn listen(self, backlog: u32) -> io::Result<TcpListener> {
-        self.inner.listen(backlog as i32).await?;
+    /// `backlog` defines the maximum number of pending connections that are
+    /// queued by the operating system at any given time. Connections are
+    /// removed from the queue with [`TcpListener::accept`]. When the queue
+    /// is full, the operating system will start rejecting connections.
+    pub async fn listen(self, backlog: i32) -> io::Result<TcpListener> {
+        self.inner.listen(backlog).await?;
         Ok(TcpListener { inner: self.inner })
     }
